+++ src/app/core/entity-components/entity-subrecord/entity-subrecord.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
-import { EntitySubrecordComponent } from "./entity-subrecord.component";
 import { KeysPipe } from "./keys-pipe/keys.pipe";
 import { MatTableModule } from "@angular/material/table";
 import { MatSortModule } from "@angular/material/sort";
@@ -15,6 +14,7 @@ import { EntityModule } from "../../entity/entity.module";
 import { AlertsModule } from "../../alerts/alerts.module";
 import { MatSnackBarModule } from "@angular/material/snack-bar";
 import { MatButtonModule } from "@angular/material/button";
+import { EntitySubrecordComponent } from "./entity-suprecord/entity-subrecord.component";
 
 @NgModule({
   declarations: [EntitySubrecordComponent, KeysPipe],
