+++ src/app/core/entity-components/entity-subrecord/entity-suprecord/entity-subrecord.component.html
