+++ src/app/core/entity-components/entity-subrecord/entity-suprecord/entity-subrecord.component.ts
@@ -1,33 +1,3 @@
-import {
-  AfterViewInit,
-  Component,
-  EventEmitter,
-  Input,
-  OnChanges,
-  OnInit,
-  Output,
-  SimpleChanges,
-  ViewChild,
-} from "@angular/core";
-import { MatSnackBar } from "@angular/material/snack-bar";
-import { MatSort } from "@angular/material/sort";
-import { MatPaginator, PageEvent } from "@angular/material/paginator";
-import { MatTableDataSource } from "@angular/material/table";
-import { ColumnDescription } from "./column-description";
-import { MediaChange, MediaObserver } from "@angular/flex-layout";
-import { FormValidationResult } from "./form-validation-result";
-import { ColumnDescriptionInputType } from "./column-description-input-type.enum";
-import { ComponentType } from "@angular/cdk/overlay";
-import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
-import { EntityMapperService } from "../../entity/entity-mapper.service";
-import { Entity } from "../../entity/entity";
-import { ShowsEntity } from "../../form-dialog/shows-entity.interface";
-import { FormDialogService } from "../../form-dialog/form-dialog.service";
-import { ConfirmationDialogService } from "../../confirmation-dialog/confirmation-dialog.service";
-import { AlertService } from "../../alerts/alert.service";
-import { DatePipe } from "@angular/common";
-import { BehaviorSubject } from "rxjs";
-
 /**
  * Generically configurable component to display and edit a list of entities in a compact way
  * that can especially be used within another entity's details view to display related entities.
@@ -41,6 +11,36 @@ import { BehaviorSubject } from "rxjs";
  * A detailed Guide on how to use this component is available:
  * - [How to display related entities]{@link /additional-documentation/how-to-guides/display-related-entities.html}
  */
+import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
+import {
+  AfterViewInit,
+  EventEmitter,
+  Component,
+  Input,
+  OnChanges,
+  OnInit,
+  Output,
+  ViewChild,
+  SimpleChanges,
+} from "@angular/core";
+import { Entity } from "../../../entity/entity";
+import { BehaviorSubject } from "rxjs";
+import { ColumnDescription } from "../column-description";
+import { FormValidationResult } from "../form-validation-result";
+import { MatTableDataSource } from "@angular/material/table";
+import { MatSort } from "@angular/material/sort";
+import { MatPaginator, PageEvent } from "@angular/material/paginator";
+import { EntityMapperService } from "../../../entity/entity-mapper.service";
+import { ConfirmationDialogService } from "../../../confirmation-dialog/confirmation-dialog.service";
+import { MatSnackBar } from "@angular/material/snack-bar";
+import { FormDialogService } from "../../../form-dialog/form-dialog.service";
+import { AlertService } from "../../../alerts/alert.service";
+import { DatePipe } from "@angular/common";
+import { MediaChange, MediaObserver } from "@angular/flex-layout";
+import { ColumnDescriptionInputType } from "../column-description-input-type.enum";
+import { ShowsEntity } from "../../../form-dialog/shows-entity.interface";
+import { ComponentType } from "@angular/cdk/overlay";
+
 @UntilDestroy()
 @Component({
   selector: "app-entity-subrecord",
